LowerDeviceStorageAccessInfo
